requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
python-multipart==0.0.6
requests==2.31.0
python-dotenv==1.0.0
numpy==1.24.3
scikit-learn==1.3.0
nltk==3.8.1 